refactor: extract write_temp_config helper and shared emit_log; bump to 0.3.6
write_temp_config:
- Create tests/unit/common.rs with a single canonical implementation
  (SystemTime nanos + AtomicU64 sequence for uniqueness under parallel tests)
- Remove the four identical per-file copies from config_validation.rs,
  gcs_storage.rs, gcs_validation.rs, and parse.rs

emit_log:
- Add src/log.rs with pub(crate) emit_log(level, ...) that handles both the
  RunEvent::Log observer path and the stderr fallback
- errors::emit and warnings::emit now both delegate to emit_log, eliminating
  the duplicated observer dispatch block

Version: 0.3.5 -> 0.3.6
