+++ crates/floe-core/src/warnings.rs
@@ -1,8 +1,4 @@
-use crate::run::events::{event_time_ms, is_observer_set, RunEvent};
-
-fn emit_stderr(message: &str) {
-    eprintln!("warn: {message}");
-}
+use crate::log::emit_log;
 
 pub fn emit(
     run_id: &str,
@@ -11,21 +7,7 @@ pub fn emit(
     code: Option<&str>,
     message: &str,
 ) {
-    if is_observer_set() {
-        let observer = crate::run::events::default_observer();
-        observer.on_event(RunEvent::Log {
-            run_id: run_id.to_string(),
-            log_level: "warn".to_string(),
-            code: code.map(ToString::to_string),
-            message: message.to_string(),
-            entity: entity.map(ToString::to_string),
-            input: input.map(ToString::to_string),
-            ts_ms: event_time_ms(),
-        });
-        return;
-    }
-
-    emit_stderr(message);
+    emit_log("warn", run_id, entity, input, code, message);
 }
 
 pub fn emit_once(
