+++ crates/floe-core/tests/unit/config/gcs_validation.rs
@@ -1,19 +1,6 @@
-use std::fs;
-use std::path::PathBuf;
-use std::time::{SystemTime, UNIX_EPOCH};
-
 use floe_core::{validate, ValidateOptions};
 
-fn write_temp_config(contents: &str) -> PathBuf {
-    let mut path = std::env::temp_dir();
-    let nanos = SystemTime::now()
-        .duration_since(UNIX_EPOCH)
-        .map(|duration| duration.as_nanos())
-        .unwrap_or(0);
-    path.push(format!("floe-config-{nanos}.yml"));
-    fs::write(&path, contents).expect("write temp config");
-    path
-}
+use super::super::common::write_temp_config;
 
 #[test]
 fn gcs_definition_requires_bucket() {
