+++ crates/floe-core/tests/unit/config/parse.rs
@@ -1,26 +1,9 @@
-use std::fs;
-use std::path::PathBuf;
-use std::sync::atomic::{AtomicU64, Ordering};
-use std::time::{SystemTime, UNIX_EPOCH};
-
 use floe_core::config::{
     IncrementalMode, SchemaEvolutionIncompatibleAction, SchemaEvolutionMode, WriteMode,
 };
 use floe_core::{load_config, validate_config_for_tests};
 
-static TEMP_CONFIG_SEQ: AtomicU64 = AtomicU64::new(0);
-
-fn write_temp_config(contents: &str) -> PathBuf {
-    let mut path = std::env::temp_dir();
-    let nanos = SystemTime::now()
-        .duration_since(UNIX_EPOCH)
-        .map(|duration| duration.as_nanos())
-        .unwrap_or(0);
-    let seq = TEMP_CONFIG_SEQ.fetch_add(1, Ordering::Relaxed);
-    path.push(format!("floe-config-{nanos}-{seq}.yml"));
-    fs::write(&path, contents).expect("write temp config");
-    path
-}
+use super::super::common::write_temp_config;
 
 #[test]
 fn parse_config_loads_report_and_defaults() {
