+++ crates/floe-core/tests/unit/mod.rs
@@ -1,3 +1,4 @@
+pub mod common;
 pub mod config;
 pub mod io;
 pub mod manifest;
